Moved navbar into components 'misc' folder
