+++ finite-automata-designer/src/app/components/misc/Navbar.tsx
