+++ finite-automata-designer/src/app/layout.tsx
@@ -2,7 +2,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Navbar from "./components/Navbar";
+import Navbar from "./components/misc/Navbar";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
